visualization/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
asyncpg==0.30.0
neo4j==5.27.0
python-dotenv==1.0.1

